+++ config/cli.php
@@ -4,6 +4,7 @@ use App\Console\Command;
 
 return function (App\Event\BuildConsoleCommands $event) {
     $event->addAliases([
+        'azuracast:acme:get-certificate' => Command\Acme\GetCertificateCommand::class,
         'azuracast:backup' => Command\Backup\BackupCommand::class,
         'azuracast:restore' => Command\Backup\RestoreCommand::class,
         'azuracast:debug:optimize-tables' => Command\Debug\OptimizeTablesCommand::class,
@@ -37,5 +38,6 @@ return function (App\Event\BuildConsoleCommands $event) {
         'queue:process' => Command\MessageQueue\ProcessCommand::class,
         'queue:clear' => Command\MessageQueue\ClearCommand::class,
         'cache:clear' => Command\ClearCacheCommand::class,
+        'acme:cert' => Command\Acme\GetCertificateCommand::class,
     ]);
 };