+++ frontend/vue/components/Admin/Settings/ServicesTab.vue
@@ -37,6 +37,49 @@
 
         <b-form-fieldset>
             <template #label>
+                <translate key="lang_section_letsencrypt">LetsEncrypt</translate>
+            </template>
+            <template #description>
+                <translate key="lang_section_letsencrypt_desc">LetsEncrypt provides simple, free SSL certificates allowing you to secure traffic through your control panel and radio streams.</translate>
+            </template>
+
+            <b-form-row>
+                <b-wrapped-form-group class="col-md-6" id="edit_form_acme_email"
+                                      :field="form.acme_email" input-type="email">
+                    <template #label="{lang}">
+                        <translate :key="lang">E-mail Address</translate>
+                    </template>
+                    <template #description="{lang}">
+                        <translate
+                            :key="lang">Enter your e-mail address to receive updates about your certificate.</translate>
+                    </template>
+                </b-wrapped-form-group>
+
+                <b-wrapped-form-group class="col-md-6" id="edit_form_acme_domains"
+                                      :field="form.acme_domains">
+                    <template #label="{lang}">
+                        <translate :key="lang">Domain Name(s)</translate>
+                    </template>
+                    <template #description="{lang}">
+                        <translate
+                            :key="lang">All listed domain names should point to this AzuraCast installation. Separate multiple domain names with commas.</translate>
+                    </template>
+                </b-wrapped-form-group>
+
+                <div class="form-group col">
+                    <b-button size="sm" variant="primary" :disabled="form.$anyDirty" @click="generateAcmeCert">
+                        <icon icon="badge"></icon>
+                        <translate key="lang_btn_acme_cert">Generate/Renew Certificate</translate>
+                        <span v-if="form.$anyDirty">
+                            (<translate key="lang_btn_acme_cert_save_changes">Save Changes first</translate>)
+                        </span>
+                    </b-button>
+                </div>
+            </b-form-row>
+        </b-form-fieldset>
+
+        <b-form-fieldset>
+            <template #label>
                 <translate key="lang_section_email_delivery">E-mail Delivery Service</translate>
             </template>
             <template #description>
@@ -177,6 +220,8 @@
             </b-form-row>
         </b-form-fieldset>
 
+        <streaming-log-modal ref="acmeModal"></streaming-log-modal>
+
         <admin-settings-test-message-modal :test-message-url="testMessageUrl"></admin-settings-test-message-modal>
     </b-tab>
 </template>
@@ -188,18 +233,25 @@ import BFormFieldset from "~/components/Form/BFormFieldset";
 import BWrappedFormCheckbox from "~/components/Form/BWrappedFormCheckbox";
 import AdminSettingsTestMessageModal from "~/components/Admin/Settings/TestMessageModal";
 import Icon from "~/components/Common/Icon";
+import StreamingLogModal from "~/components/Common/StreamingLogModal";
 
 export default {
     name: 'SettingsServicesTab',
     components: {
+        StreamingLogModal,
         Icon,
-        AdminSettingsTestMessageModal, BWrappedFormCheckbox, BFormFieldset, BWrappedFormGroup, BFormMarkup
+        AdminSettingsTestMessageModal,
+        BWrappedFormCheckbox,
+        BFormFieldset,
+        BWrappedFormGroup,
+        BFormMarkup
     },
     props: {
         form: Object,
         tabClass: {},
         releaseChannel: String,
         testMessageUrl: String,
+        acmeUrl: String,
     },
     computed: {
         langTabTitle() {
@@ -226,6 +278,15 @@ export default {
                 }
             ]
         },
+    },
+    methods: {
+        generateAcmeCert() {
+            this.$wrapWithLoading(
+                this.axios.put(this.acmeUrl)
+            ).then((resp) => {
+                this.$refs.acmeModal.show(resp.data.links.log);
+            });
+        }
     }
 }
 </script>